abstracted challenge biz logic to service layer
